Renamed constrains. Validation on delete categories.

+++ Application/Categories/Commands/Delete/DeleteCategoryCommandHandler.cs
@@ -3,7 +3,7 @@ using Application.Common.Interfaces;
 using Domain.Entities;
 using Gallery.Common.Helpers;
 using MediatR;
-using System;
+using Microsoft.EntityFrameworkCore;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
@@ -21,32 +21,22 @@ namespace Application.Categories.Commands.Delete
 
         public async Task<Unit> Handle(DeleteCategoryCommand command, CancellationToken cancellationToken)
         {
-            string errorMessage = null;
-
             if (command.Id <= 0)
                 throw new HttpStatusCodeException(HttpStatusCode.BadRequest, ErrorMessages.IdCanNotBeLowerThanOne);
+
+            bool exists = await _context.Pictures.AnyAsync(x => x.CategoryId == command.Id);
+
+            if (exists)
+                throw new HttpStatusCodeException(HttpStatusCode.NotFound, ErrorMessages.CanNotDeleteCategory);
 
             Category entity = await _context.Categories.FindAsync(command.Id);
 
             if (entity == null)
                 throw new HttpStatusCodeException(HttpStatusCode.NotFound, ErrorMessages.CategoryNotFound);
 
+            _context.Categories.Remove(entity);
 
-            try
-            {
-                _context.Categories.Remove(entity);
-
-                await _context.SaveChangesAsync(cancellationToken);
-            }
-            catch (Exception exception)
-            {
-                if (exception.Message.Contains("The DELETE statement conflicted with the REFERENCE"))
-                    errorMessage = ErrorMessages.CanNotDeleteCategory;
-                else
-                    errorMessage = exception.Message;
-
-                throw new HttpStatusCodeException(HttpStatusCode.BadRequest, errorMessage);
-            }
+            await _context.SaveChangesAsync(cancellationToken);
 
             return Unit.Value;
         }
